requirement.txt
google-cloud-bigquery
pyyaml
slackweb